Atualizando sem função

+++ index.js
@@ -7,27 +7,6 @@ const port = 3000;
 
 app.use(express.json());
 
-const validaId = (res, id) => {
-  if (!mongoose.Types.ObjectId.isValid(id)) {
-    res.status(422).send({ error: "ID inválido" })
-    return;
-  }
-};
-
-const jogoExiste = (res, jogo) => {
-  if (!jogo) {
-    res.status(404).send({ erro: "Jogo não encontrado!" })
-    return;
-  }
-};
-
-const validaJogo = (res, jogo) => {
-  if (!jogo || !jogo.nome || !jogo.imagem) {
-    res.status(400).send({ error: "Jogo inválido!" })
-    return;
-  }
-};
-
 app.get("/", (req, res) => {
   res.send({ info: "Projeto de jogos com MongoDB" });
 });
@@ -40,11 +19,17 @@ app.get("/jogos", async (req, res) => {
 app.get("/jogos/:id", async (req, res) => {
   const id = req.params.id;
 
-  validaId(res, id);
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    res.status(422).send({ error: "ID inválido" });
+    return;
+  }
 
   const jogo = await jogoSchema.findById(id);
 
-  jogoExiste(res, jogo);
+  if (!jogo) {
+    res.status(404).send({ erro: "Jogo não encontrado!" });
+    return;
+  }
 
   res.send({ jogo });
 });
@@ -52,7 +37,10 @@ app.get("/jogos/:id", async (req, res) => {
 app.post("/jogos", async (req, res) => {
   const jogo = req.body;
 
-  validaJogo(res, jogo);
+  if (!jogo) {
+    res.status(404).send({ erro: "Jogo não encontrado!" });
+    return;
+  }
 
   const novoJogo = await new jogoSchema(jogo).save();
   res.status(201).send({ novoJogo });
@@ -61,30 +49,44 @@ app.post("/jogos", async (req, res) => {
 app.put("/jogos/:id", async (req, res) => {
   const id = req.params.id;
 
-  validaId(res, id);
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    res.status(422).send({ error: "ID inválido" });
+    return;
+  }
 
   const jogo = await jogoSchema.findById(id);
 
-  jogoExiste(res, jogo);
+  if (!jogo) {
+    res.status(404).send({ erro: "Jogo não encontrado!" });
+    return;
+  }
 
   const novoJogo = req.body;
 
-  validaJogo(res, jogo);
+  if (!jogo || !jogo.nome || !jogo.imagem) {
+    res.status(400).send({ error: "Jogo inválido!" });
+    return;
+  }
 
   await jogoSchema.findOneAndUpdate({ _id: id }, novoJogo);
   const jogoAtualizado = await jogoSchema.findById(id);
 
-  res.send({ jogoAtualizado });
 });
 
 app.delete("/jogos/:id", async (req, res) => {
   const id = req.params.id;
 
-  validaId(res, id);
-
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    res.status(422).send({ error: "ID inválido" });
+    return;
+  }
+  
   const jogo = await jogoSchema.findById(id);
 
-  jogoExiste(res, jogo);
+  if (!jogo) {
+    res.status(404).send({ erro: "Jogo não encontrado!" });
+    return;
+  }
 
   await jogoSchema.findByIdAndDelete(id);
   res.send({ message: "Jogo excluído com sucesso!" });
